HU-20
C2 Display assigned permissions in frontend table
