+++ backend/routes/userRoutes.js
@@ -123,81 +123,6 @@ router.delete("/:id", async (req, res) => {
     }
 });
 
-// PATCH: Asignar rol y permisos
-router.patch('/:id/role', async (req, res) => {
-  const { id } = req.params;
-  const { isAdmin, permissions } = req.body;
-
-  try {
-    // Validar que isAdmin sea booleano
-    if (typeof isAdmin !== 'boolean') {
-      return res.status(400).json({ message: 'El campo "isAdmin" debe ser booleano.' });
-    }
-
-    // Si isAdmin es true, validar que haya al menos un permiso
-    if (isAdmin && (!Array.isArray(permissions) || permissions.length === 0)) {
-      return res.status(400).json({ message: 'Debes asignar al menos un permiso al administrador.' });
-    }
-
-    // Actualizar el campo isAdmin en el modelo User
-    const user = await User.findByIdAndUpdate(id, { isAdmin }, { new: true });
-    if (!user) return res.status(404).json({ message: 'Usuario no encontrado.' });
-
-    if (isAdmin) {
-      // Crear o actualizar el documento Admin con los permisos
-      const admin = await Admin.findOneAndUpdate(
-        { user: id },
-        { permisos: permissions },
-        { upsert: true, new: true }
-      );
-      return res.status(200).json({ user, admin });
-    } else {
-      // Si se revoca el rol de administrador, eliminar permisos asociados
-      await Admin.deleteOne({ user: id });
-      return res.status(200).json({ user });
-    }
-  } catch (error) {
-    console.error('Error actualizando rol:', error);
-    return res.status(500).json({ message: 'Error interno del servidor.' });
-  }
-});
-
-// GET /api/users — Trae todos los usuarios con sus permisos si son admin
-router.get('/', async (req, res) => {
-  try {
-    // Traemos todos los usuarios
-    const users = await User.find({}, '_id name email isAdmin');
-
-    // Creamos un array de sus _id
-    const userIds = users.map(user => user._id);
-
-    // Buscamos todos los documentos Admin asociados
-    const adminDocs = await Admin.find({ user: { $in: userIds } });
-
-    // Creamos un mapa rápido {userId: permisos[]}
-    const permisosMap = {};
-    adminDocs.forEach(doc => {
-      permisosMap[doc.user.toString()] = doc.permisos;
-    });
-
-    // Componemos el resultado final con los permisos solo si isAdmin === true
-    const usersWithPermisos = users.map(user => ({
-      _id: user._id,
-      name: user.name,
-      email: user.email,
-      isAdmin: user.isAdmin,
-      permissions: user.isAdmin
-        ? (permisosMap[user._id.toString()] || [])
-        : []
-    }));
-
-    return res.status(200).json(usersWithPermisos);
-
-  } catch (error) {
-    console.error('Error al obtener usuarios con permisos:', error);
-    return res.status(500).json({ message: 'Error interno del servidor' });
-  }
-});
